+++ network/p2p/p2p.go
@@ -1,6 +1,7 @@
 package p2pv1
 
 import (
+	"bytes"
 	"context"
 	"github.com/bloxapp/ssv/network"
 	"github.com/bloxapp/ssv/network/discovery"
@@ -8,6 +9,7 @@ import (
 	forksfactory "github.com/bloxapp/ssv/network/forks/factory"
 	"github.com/bloxapp/ssv/network/peers"
 	"github.com/bloxapp/ssv/network/peers/connections"
+	"github.com/bloxapp/ssv/network/records"
 	"github.com/bloxapp/ssv/network/streams"
 	"github.com/bloxapp/ssv/network/topics"
 	"github.com/bloxapp/ssv/utils/async"
@@ -64,7 +66,7 @@ type p2pNetwork struct {
 
 	backoffConnector *libp2pdisc.BackoffConnector
 	subnets          []byte
-	connManager      connmgrcore.ConnManager
+	libConnManager   connmgrcore.ConnManager
 }
 
 // New creates a new p2p network
@@ -95,7 +97,7 @@ func (n *p2pNetwork) Close() error {
 	atomic.SwapInt32(&n.state, stateClosing)
 	defer atomic.StoreInt32(&n.state, stateClosed)
 	n.cancel()
-	if err := n.connManager.Close(); err != nil {
+	if err := n.libConnManager.Close(); err != nil {
 		n.logger.Warn("could not close discovery", zap.Error(err))
 	}
 	if err := n.disc.Close(); err != nil {
@@ -122,15 +124,18 @@ func (n *p2pNetwork) Start() error {
 	go n.startDiscovery()
 
 	async.Interval(n.ctx, connManagerGCInterval, func() {
-		currentCount := len(n.host.Peerstore().Peers())
+		allPeers := n.host.Peerstore().Peers()
+		currentCount := len(allPeers)
 		if currentCount > n.cfg.MaxPeers {
 			return
 		}
 		ctx, cancel := context.WithCancel(n.ctx)
 		defer cancel()
 
-		n.tagBestPeers(currentCount - 1)
-		n.trimPeers(ctx)
+		connMgr := peers.NewConnManager(n.logger, n.libConnManager, n.idx)
+		mySubnets := records.Subnets(n.subnets).Clone()
+		connMgr.TagBestPeers(currentCount-1, mySubnets, allPeers, n.cfg.TopicMaxPeers)
+		connMgr.TrimPeers(ctx, n.host.Network())
 	})
 
 	async.Interval(n.ctx, peerIndexGCInterval, func() {
@@ -203,4 +208,68 @@ func (n *p2pNetwork) startDiscovery() {
 
 func (n *p2pNetwork) isReady() bool {
 	return atomic.LoadInt32(&n.state) == stateReady
+}
+
+// UpdateSubnets will update the registered subnets according to active validators
+// NOTE: it won't subscribe to the subnets (use subscribeToSubnets for that)
+func (n *p2pNetwork) UpdateSubnets() {
+	visited := make(map[int]bool)
+	n.activeValidatorsLock.Lock()
+	last := make([]byte, len(n.subnets))
+	if len(n.subnets) > 0 {
+		copy(last, n.subnets)
+	}
+	newSubnets := make([]byte, n.fork.Subnets())
+	for pkHex, state := range n.activeValidators {
+		if state == validatorStateInactive {
+			continue
+		}
+		subnet := n.fork.ValidatorSubnet(pkHex)
+		if _, ok := visited[subnet]; ok {
+			continue
+		}
+		newSubnets[subnet] = byte(1)
+	}
+	subnetsToAdd := make([]int, 0)
+	if !bytes.Equal(newSubnets, last) { // have changes
+		n.subnets = newSubnets
+		for i, b := range newSubnets {
+			if b == byte(1) {
+				subnetsToAdd = append(subnetsToAdd, i)
+			}
+		}
+	}
+	n.activeValidatorsLock.Unlock()
+
+	if len(subnetsToAdd) == 0 {
+		n.logger.Debug("no changes in subnets")
+		return
+	}
+
+	self := n.idx.Self()
+	self.Metadata.Subnets = records.Subnets(n.subnets).String()
+	n.idx.UpdateSelfRecord(self)
+
+	allSubs, _ := records.Subnets{}.FromString(records.AllSubnets)
+	subnetsList := records.SharedSubnets(allSubs, n.subnets, 0)
+	n.logger.Debug("updated subnets (node-info)", zap.Any("subnets", subnetsList))
+
+	err := n.disc.RegisterSubnets(subnetsToAdd...)
+	if err != nil {
+		n.logger.Warn("could not register subnets", zap.Error(err))
+		return
+	}
+	n.logger.Debug("updated subnets (discovery)", zap.Any("subnets", n.subnets))
+}
+
+// getMaxPeers returns max peers of the given topic.
+func (n *p2pNetwork) getMaxPeers(topic string) int {
+	if len(topic) == 0 {
+		return n.cfg.MaxPeers
+	}
+	baseName := n.fork.GetTopicBaseName(topic)
+	if baseName == n.fork.DecidedTopic() { // allow more peers for decided topic
+		return n.cfg.TopicMaxPeers * 2
+	}
+	return n.cfg.TopicMaxPeers
 }
