+++ network/peers/scores.go
@@ -2,6 +2,7 @@ package peers
 
 import (
 	"github.com/libp2p/go-libp2p-core/peer"
+	"sort"
 	"sync"
 )
 
@@ -61,3 +62,33 @@ wantedScoresLoop:
 	}
 	return scores, nil
 }
+
+// GetTopScores accepts a map of scores and returns the best n peers
+func GetTopScores(peerScores map[peer.ID]int, n int) map[peer.ID]int {
+	pl := make(peerScoresList, len(peerScores))
+	i := 0
+	for k, v := range peerScores {
+		pl[i] = peerScorePair{k, v}
+		i++
+	}
+	sort.Sort(sort.Reverse(pl))
+	res := make(map[peer.ID]int)
+	for _, item := range pl {
+		res[item.Key] = item.Value
+		if len(res) >= n {
+			break
+		}
+	}
+	return res
+}
+
+type peerScorePair struct {
+	Key   peer.ID
+	Value int
+}
+
+type peerScoresList []peerScorePair
+
+func (p peerScoresList) Len() int           { return len(p) }
+func (p peerScoresList) Less(i, j int) bool { return p[i].Value < p[j].Value }
+func (p peerScoresList) Swap(i, j int)      { p[i], p[j] = p[j], p[i] }
