+++ network/peers/subnets_test.go
@@ -6,6 +6,7 @@ import (
 	"github.com/libp2p/go-libp2p-core/crypto"
 	"github.com/libp2p/go-libp2p-core/peer"
 	"github.com/stretchr/testify/require"
+	"math/rand"
 	"testing"
 )
 
@@ -51,4 +52,37 @@ func TestSubnetsIndex(t *testing.T) {
 
 	require.Len(t, subnetsIdx.GetSubnetPeers(0), 0)
 	require.Len(t, subnetsIdx.GetSubnetPeers(10), 0)
+}
+
+func TestSubnetsDistributionScores(t *testing.T) {
+	nsubnets := 128
+	mysubnets := make(records.Subnets, nsubnets)
+	allSubs, _ := records.Subnets{}.FromString(records.AllSubnets)
+	for sub := range allSubs {
+		if sub%2 == 0 {
+			mysubnets[sub] = byte(0)
+		} else {
+			mysubnets[sub] = byte(1)
+		}
+	}
+	stats := &SubnetsStats{
+		PeersCount: make([]int, len(mysubnets)),
+		Connected:  make([]int, len(mysubnets)),
+	}
+	for sub := range mysubnets {
+		stats.Connected[sub] = 1 + rand.Intn(20)
+		stats.PeersCount[sub] = stats.Connected[sub] + rand.Intn(10)
+	}
+	stats.Connected[1] = 0
+	stats.Connected[3] = 1
+	stats.Connected[5] = 30
+	stats.PeersCount[5] = 30
+
+	distScores := GetSubnetsDistributionScores(stats, 3, mysubnets, 5)
+
+	require.Len(t, distScores, len(mysubnets))
+	require.Equal(t, 0, distScores[0])
+	require.Equal(t, 2, distScores[1])
+	require.Equal(t, 1, distScores[3])
+	require.Equal(t, -1, distScores[5])
 }
